Show informative labels in my investment if funds raising phase not finished yet
It's not possible to lock or unlock shares before raising funds is finished. This has technical and UX/legal reasons:
- Technical: If it's possible to unlock, transfer is possible, and we can't ensure that a user doesn't buy more than upper limit.
- UX/legal: Trading shares before the project is successfully funded seems dubious. If the funding, and thus the project, fails, what is being traded? seems an invitation for scams and legal issues.

Note that, teal currently allows to *lock* shares (there doesn't seem to be a reason not to), but given that this is irrelevant to investors (if they can't unlock, there's nothing to lock), we disable both to prevent confusion.